requirements.txt
accelerate==1.8.1
acres==0.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
arrow==1.3.0
attrs==25.3.0
bertopic==0.17.3
blinker==1.9.0
blis==1.3.0
cachetools==5.5.2
camelot-py==1.0.9
catalogue==2.0.10
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.1.8
cloudpathlib==0.21.1
colorama==0.4.6
confection==0.1.5
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.2
cryptography==45.0.6
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
diff-match-patch==20230430
distro==1.9.0
dnspython==2.7.0
docx==0.2.4
dotenv==0.9.9
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
etelemetry==0.3.1
faiss-cpu==1.12.0
Faker==25.9.2
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
fonttools==4.58.4
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.5.1
generativeai==0.0.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google_search_results==2.4.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
griffe==0.36.9
grpcio==1.73.1
grpcio-status==1.71.2
guardrail==0.1.1
guardrails-ai==0.6.6
guardrails-api-client==0.4.0a1
guardrails_hub_types==0.0.4
h11==0.16.0
hdbscan==0.8.40
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.2
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
isoduration==20.11.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jpype1==1.6.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
keybert==0.9.0
kiwisolver==1.4.8
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-text-splitters==0.3.9
langcodes==3.5.0
langsmith==0.4.8
language_data==1.3.0
lark==1.2.2
litellm==1.74.6
llvmlite==0.44.0
loguru==0.7.3
looseversion==1.3.0
lxml==4.9.4
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.3
murmurhash==1.0.13
mypy_extensions==1.1.0
narwhals==1.47.0
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
nltk==3.9.1
numba==0.61.2
numpy==2.2.6
openai==1.97.0
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.35.0
opentelemetry-exporter-otlp-proto-http==1.35.0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
orjson==3.11.0
packaging==24.2
pandas==2.3.0
pathlib==1.0.1
pdfminer.six==20250506
pillow==11.3.0
pinecone==7.3.0
pinecone-plugin-assistant==1.7.0
pinecone-plugin-interface==0.0.7
plotly==6.2.0
preshed==3.0.10
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
prov==2.1.1
psutil==7.0.0
psycopg2-binary==2.9.10
puremagic==1.29
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydash==7.0.7
pydot==4.0.1
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.14.1
PyMuPDF==1.26.1
pynndescent==0.5.13
pyparsing==3.2.3
pypdf==3.17.4
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
RapidFuzz==3.13.0
rdflib==7.1.4
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==13.9.4
rpds-py==0.26.0
rsa==4.9.1
rstr==3.2.2
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
semver==3.0.4
sentence-transformers==5.0.0
sentencepiece==0.2.0
serpapi==0.1.5
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.43
srsly==2.5.1
sympy==1.14.0
tabula-py==2.10.0
tabulate==0.9.0
tenacity==9.1.2
textblob==0.19.0
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
traits==7.0.2
transformers==4.53.1
typer==0.15.4
types-python-dateutil==2.9.0.20250708
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
umap-learn==0.5.9.post2
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.0.7
vaderSentiment==3.3.2
wasabi==1.1.3
weasel==0.4.1
webcolors==24.11.1
Werkzeug==3.1.3
win32_setctime==1.2.0
wrapt==1.17.2
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0